+++ src/main/java/pro/sky/hw27/controller/DepartmentController.java
@@ -5,6 +5,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import pro.sky.hw27.Employee;
+import pro.sky.hw27.employeeService.DepartmentService;
 import pro.sky.hw27.employeeService.EmployeeService;
 
 import java.util.List;
@@ -13,30 +14,35 @@ import java.util.Map;
 @RestController
 @RequestMapping("/department")
 public class DepartmentController {
-    public final EmployeeService employeeService;
+    public final DepartmentService departmentService;
 
-    public DepartmentController(EmployeeService employeeService) {
-        this.employeeService = employeeService;
+    public DepartmentController(DepartmentService departmentService) {
+        this.departmentService = departmentService;
+    }
+    @GetMapping("/add")
+    public Employee addEmployee(@RequestParam("name") String name,@RequestParam("id") int id,@RequestParam("salary") int salary) {
+        return departmentService.addEmployee(name,id,salary);
+
     }
 
     @GetMapping("/all-by-dep")
     public List<Employee> getAllByDep(@RequestParam("id") int id) {
-        return employeeService.getAllByDep(id);
+        return departmentService.getAllByDep(id);
     }
 
     @GetMapping("/all")
     public Map<Integer,List<Employee>> getAll() {
-        return employeeService.getAll();
+        return departmentService.getAll();
     }
 
     @GetMapping("/min-salary")
     public Employee minSalary(@RequestParam("id") int id ) {
-        return employeeService.minSalary(id);
+        return departmentService.minSalary(id);
     }
 
     @GetMapping("/max-salary")
     public Employee maxSalary(@RequestParam("id") int id ) {
-        return employeeService.maxSalary(id);
+        return departmentService.maxSalary(id);
 
     }
 }
