+++ src/main/java/pro/sky/hw27/controller/EmployeeController.java
@@ -32,10 +32,5 @@ public class EmployeeController {
 
     }
 
-    @GetMapping("/add")
-    public Employee addEmployee(@RequestParam("name") String name,@RequestParam("id") int id,@RequestParam("salary") int salary) {
-        return employeeService.addEmployee(name,id,salary);
-
-    }
 
 }
